handle for stable vs unstable onPluginInit
